Move media reference DTO to query layer
