+++ src/server/packages/Media/Application/Query/MediaDetailQueryServiceInterface.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Media\Application\Query;
 
 use Media\Domain\Models\MediaId;
-use Media\Domain\Models\MediaReferencedSong;
 
 interface MediaDetailQueryServiceInterface
 {
