+++ src/server/packages/Media/Application/Query/MediaReferencedSong.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Media\Application\Query;
+
+readonly class MediaReferencedSong
+{
+    public function __construct(
+        public string $songId,
+        public string $title,
+        public int $songOrderNo,
+        public int $mediaOrderNo,
+    ) {
+    }
+}
