+++ src/server/packages/Media/Application/UseCase/Get/GetOutputData.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Media\Application\UseCase\Get;
 
+use Media\Application\Query\MediaReferencedSong;
 use Media\Domain\Models\Media;
-use Media\Domain\Models\MediaReferencedSong;
 
 readonly class GetOutputData
 {
