+++ src/server/packages/Media/Infrastructures/MediaDetailQueryService.php
@@ -6,8 +6,8 @@ namespace Media\Infrastructures;
 
 use App\Models\Song\SongMediaLink as ModelsSongMediaLink;
 use Media\Application\Query\MediaDetailQueryServiceInterface;
+use Media\Application\Query\MediaReferencedSong;
 use Media\Domain\Models\MediaId;
-use Media\Domain\Models\MediaReferencedSong;
 use Override;
 use Support\Contracts\Uuid\UuidConverterInterface;
 
@@ -29,7 +29,7 @@ readonly class MediaDetailQueryService implements MediaDetailQueryServiceInterfa
                     fn (ModelsSongMediaLink $link): int => $link->song->order_no,
                     fn (ModelsSongMediaLink $link): int => $link->order_no,
                 ])
-                ->map(fn (ModelsSongMediaLink $link): MediaReferencedSong => MediaReferencedSong::reconstruct(
+                ->map(fn (ModelsSongMediaLink $link): MediaReferencedSong => new MediaReferencedSong(
                     $this->converter->toUuid($link->song_id),
                     $link->song->title,
                     $link->song->order_no,
